Desktop/Pneumonia_1/requirements.txt
Flask
Gunicorn
h5py
Jinja2
keras
Keras-Applications
Keras-Preprocessing
tensorflow
tensorflow-estimator
MarkupSafe
protobuf
Werkzeug
numpy		
Pillow